// medicare-ai-production/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as appointments from "../appointments.js";
import type * as chat from "../chat.js";
import type * as consultations from "../consultations.js";
import type * as emergency from "../emergency.js";
import type * as medications from "../medications.js";
import type * as notifications from "../notifications.js";
import type * as patients from "../patients.js";
import type * as prescriptions from "../prescriptions.js";
import type * as reports from "../reports.js";
import type * as seed from "../seed.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  appointments: typeof appointments;
  chat: typeof chat;
  consultations: typeof consultations;
  emergency: typeof emergency;
  medications: typeof medications;
  notifications: typeof notifications;
  patients: typeof patients;
  prescriptions: typeof prescriptions;
  reports: typeof reports;
  seed: typeof seed;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};